corrected script name issue

+++ setup-buildenv.sh
@@ -1,23 +1,27 @@
 #!/bin/sh
 
-./setup-deviceenv.sh
+./setup-installenv.sh
 
-mkdir bolos_env
+if [ -d bolos_env ]; then
 
-cd bolos_env
+  mkdir bolos_env
 
-wget https://launchpad.net/gcc-arm-embedded/5.0/5-2016-q3-update/+download/gcc-arm-none-eabi-5_4-2016q3-20160926-linux.tar.bz2
-bunzip2 gcc-arm-none-eabi-5_4-2016q3-20160926-linux.tar.bz2
-tar xvf gcc-arm-none-eabi-5_4-2016q3-20160926-linux.tar
-rm -f gcc-arm-none-eabi-5_4-2016q3-20160926-linux.tar
+  cd bolos_env
 
-wget http://releases.llvm.org/4.0.0/clang+llvm-4.0.0-x86_64-linux-gnu-ubuntu-16.10.tar.xz
-tar xvf clang+llvm-4.0.0-x86_64-linux-gnu-ubuntu-16.10.tar.xz
-rm -f clang+llvm-4.0.0-x86_64-linux-gnu-ubuntu-16.10.tar.xz
+  wget https://launchpad.net/gcc-arm-embedded/5.0/5-2016-q3-update/+download/gcc-arm-none-eabi-5_4-2016q3-20160926-linux.tar.bz2
+  bunzip2 gcc-arm-none-eabi-5_4-2016q3-20160926-linux.tar.bz2
+  tar xvf gcc-arm-none-eabi-5_4-2016q3-20160926-linux.tar
+  rm -f gcc-arm-none-eabi-5_4-2016q3-20160926-linux.tar
 
-ln -sf clang+llvm-4.0.0-x86_64-linux-gnu-ubuntu-16.10 clang-arm-fropi 
+  wget http://releases.llvm.org/4.0.0/clang+llvm-4.0.0-x86_64-linux-gnu-ubuntu-16.10.tar.xz
+  tar xvf clang+llvm-4.0.0-x86_64-linux-gnu-ubuntu-16.10.tar.xz
+  rm -f clang+llvm-4.0.0-x86_64-linux-gnu-ubuntu-16.10.tar.xz
 
-cd ..
+  ln -sf clang+llvm-4.0.0-x86_64-linux-gnu-ubuntu-16.10 clang-arm-fropi 
+
+  cd ..
+
+fi
 
 
 . ./configure.env
